app/src/main/java/com/example/haekalmoviecatalogue/ui/detail/moviedetail/DetailMovieViewModel.kt
package com.example.haekalmoviecatalogue.ui.detail.moviedetail

import androidx.lifecycle.ViewModel
import com.example.haekalmoviecatalogue.data.MovieEntity
import com.example.haekalmoviecatalogue.utils.DataDummy

class DetailMovieViewModel : ViewModel() {
    private lateinit var movieId: String

    fun setSelectedMovie(movieId: String) {
        this.movieId = movieId
    }

    fun getMovie(): MovieEntity {
        lateinit var movie: MovieEntity
        val movieEntities = DataDummy.generateDummyMovies()
        for (movieEntity in movieEntities) {
            if (movieEntity.movieId == movieId) {
                movie = movieEntity
            }
        }
        return movie
    }
}